rebased and resolved conflicts
